Refactor -- more misc. corrections.
- Fixed some conversions.
- Fixed some locals that hide previous declarations.
- Fixed unreachable code in mob::delete_old_status_effects().

+++ Source/source/controls.cpp
@@ -598,7 +598,7 @@ void gameplay::handle_button(
             bool finished = false;
             do {
                 next_maturity =
-                    sum_and_wrap(
+                    (size_t) sum_and_wrap(
                         next_maturity,
                         (button == BUTTON_NEXT_MATURITY ? 1 : -1),
                         N_MATURITIES
