+++ Source/source/vars.h
@@ -366,7 +366,7 @@ extern int scr_h;
 extern int scr_w;
 extern ALLEGRO_TRANSFORM screen_to_world_transform;
 extern sector_types_manager sector_types;
-extern unsigned int selected_spray;
+extern size_t selected_spray;
 extern replay session_replay;
 extern unsigned char ship_beam_ring_color[3];
 extern bool ship_beam_ring_color_up[3];
